+++ Math/NumberTheory/Quadratic/EisensteinIntegers.hs
@@ -38,7 +38,7 @@ import Math.NumberTheory.Moduli.Sqrt
 import qualified Math.NumberTheory.Primes.Sieve         as Sieve
 import qualified Math.NumberTheory.Primes.Testing       as Testing
 import Math.NumberTheory.Primes.Types
-import qualified Math.NumberTheory.UniqueFactorisation  as U
+import qualified Math.NumberTheory.Primes  as U
 import Math.NumberTheory.Utils                          (mergeBy)
 import Math.NumberTheory.Utils.FromIntegral
